REFACTOR enable all cops by default
Currently, cops are disabled by default and only
ones that are explicitely enabled will be used.
Well, this is fine because it gave us more control
but a pain to maintain because we will need to
manually add new ones. Let's change this approach
by enabling all by default and disabling ones we
don't want !

Details:
- ENABLE all cops
- UPDATE `.rubocop.yml`
- UPDATE documentation